tests/wpt/tests/html/dom/partial-updates/tentative/stream-append-has-invalidation.html

<!DOCTYPE html>
<meta charset="utf-8" />
<title>HTML partial updates - :has invalidation with streaming</title>
<link rel="help" href="https://github.com/WICG/declarative-partial-updates" />
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<style>
  .container {
    width: 200px;
    height: 100px;
    background: green;
  }

  .container:has(.red) {
    background: red;
  }

  .container:has(.blue) {
    background: blue;
  }
</style>
<div class="container" marker="content">
  <?start name=content><span class="red">Has red</span><?end name=content>
</div>

<script>
  const container = document.querySelector(".container");
  async function update(html) {
    const writer = container.streamAppendHTMLUnsafe({runScripts: true}).getWriter();
    await writer.write(`
      <template for=content>${html}</template>`);
    await writer.close();
  }
  promise_test(async () => {
    for (let i = 0; i < 20; ++i) {
      await update('<?start name=content><span class="blue">Has blue</span>');
      await new Promise((resolve) => requestAnimationFrame(resolve));
      await update("Green (no span)");
      await new Promise((resolve) => requestAnimationFrame(resolve));
      assert_equals(container.textContent.trim(), "Green (no span)", container.innerHTML);
      assert_equals(
        getComputedStyle(container).backgroundColor,
        "rgb(0, 128, 0)"
      );
    }
  }, "");
</script>
